adds updatedUserHydration function
